refactor(embeds): split the files the catalog additions pushed over the limit
The 350-line gate caught four. Each splits along a seam it already had rather
than at an arbitrary line:

- `document_cards.rs` — note.com, Figma, and Google Slides are all cards for a
  document rather than a post or a video, and none of them belongs with the
  social ones.
- `catalog_tests.rs` — the #1100 additions and the rules they share.
- `embed_tag_bindings.rs` — the tag list is a separate question from the
  transform that consumes it.
- `media-marker.ts` — the pre-scan, with the drift it now prevents.

`article_card` becomes visible to the sibling module that took note.com.
